third_party/WebKit/LayoutTests/editing/style/create-block-for-style-002.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<p>
Should see "bar" centered in its own red box, with a left-aligned "foo" line above and a left-aligned "baz" line below.
</p>
<div id="sample" contenteditable>foo<br>bar<br>baz</div>
<div id="log"></div>
<script>
test(function() {
    var selection = window.getSelection();
    var sample = document.getElementById('sample');

    selection.collapse(sample, 2);
    document.execCommand('justifyCenter');

    assert_equals(sample.innerHTML, 'foo<br><div style="text-align: center;">bar</div>baz', 'innerHTML');
    assert_true(selection.isCollapsed, 'isCollapsed');
    assert_equals(selection.anchorNode, sample.childNodes[2].firstChild, 'anchorNode');
    assert_equals(selection.anchorOffset, 0, 'anchorOffset');
});
</script>
